backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
git+https://github.com/openai/CLIP.git
faiss-cpu>=1.7.4
numpy>=1.24.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.4.0
aiofiles>=23.0.0
httpx>=0.25.0
scikit-learn>=1.3.0
tqdm>=4.65.0
opencv-python>=4.8.0
aiohttp>=3.9.0
diskcache>=5.6.3
imagehash>=4.3.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0